Delay multiple Tasks
As a second step toward executing status updates in-bulk, mark a group
of selected `Task` records as "delayed" by marking their new
`delayed_at` DATETIME column to "now".

When a `Task` is either "complete" or "delayed", exclude it from the
"Todo" section of the list.

To submit the shared `<form>` element to a different URL, declare the
"Delay" [`<button>` element with the `formaction`
attribute][mdn-formaction] set to the `delays_path`.

[mdn-formaction]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/button#attr-formaction