Open the `<dialog>` with a `<button>`
Nest the page's content within a [`<main>` element][mdn-main], and
render a [`<dialog>` element][mdn-dialog] as its sibling, which we're
polyfilling with [Google's `dialog-polyfill` package][dialog-polyfill].

Move the `Task` selection elements from the `tasks#index` template to a
new `events#new` template, and render it in-line within the `<dialog>`
element.

In its place, render a set of `<button>` elements that will open the
`<dialog>` to further manage Bulk Action selection.

In order to manage that [`<dialog>` element][mdn-dialog] interaction,
this commit introduces the `ControlsDialog` and `Dialog` Stimulus
controllers to work in tandem.

[mdn-main]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/main
[mdn-dialog]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/dialog
[dialog-polyfill]: https://github.com/GoogleChrome/dialog-polyfill