Mark <main inert> when <dialog> open
When showing the `<dialog>` element modally, prevent tab focus and other
interaction with the rest of the page by marking the `<main>` element
with the [`[inert]` attribute][mdn-inert]. This commit includes [an
`[inert]` attribute polyfill][wicg-inert].

To ensure that the application is tab-navigable, add a System Test to
exercise that behavior. To make assertions about focus state, add a
dependency on the [`capybara_accessible_selectors`
gem][capybara_accessible_selectors] to augment the standard set of
[`Capybara` selectors and filters][capybara-selectors].

Finally, as part of pushing the [`dialog-polyfill`][dialog-polyfill],
include [an accessibility-motivated polyfill][dialog-focus-restore] to
manage focus state when a `<dialog>` is opened and closed.

[wicg-inert]: https://github.com/WICG/inert/blob/master/explainer.md
[mdn-inert]: https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/inert
[capybara_accessible_selectors]: https://github.com/citizensadvice/capybara_accessible_selectors
[capybara-selectors]: https://www.rubydoc.info/github/teamcapybara/capybara/Capybara/Selector
[dialog-polyfill]: https://github.com/GoogleChrome/dialog-polyfill
[dialog-focus-restore]: https://gist.github.com/samthor/babe9fad4a65625b301ba482dad284d1