Treat Lists as separate tab stops
Declare the `TabstopList` Stimulus controller, along with the
`trapFocus()` and `navigate()` actions.

When a list gains focus, find all focusable elements (like `<button>`,
`<input>`, `<select>`, etc), and prevent focus by [setting their
`[tabindex="-1"]`][mdn-tabindex-1]. Then, re-set the focused element's
`[tabindex="0"]` to treat it as the place to resume once focus is
regained.

To go along with that restriction, treat the up and down arrows as
keyboard shortcuts to move between focusable items in the list, and add
a System Test to exercise that behavior.

To make assertions about focus state, add a dependency on the
[`capybara_accessible_selectors` gem][capybara_accessible_selectors] to
augment the standard set of [`Capybara` selectors and
filters][capybara-selectors].

[mdn-tabindex]: https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/tabindex
[capybara_accessible_selectors]: https://github.com/citizensadvice/capybara_accessible_selectors
[capybara-selectors]: https://www.rubydoc.info/github/teamcapybara/capybara/Capybara/Selector