Treat Lists as separate tab stops
Declare the `TabstopList` Stimulus controller, along with the
`trapFocus()` and `navigate()` actions.

When a list gains focus, find all focusable elements (like `<button>`,
`<input>`, `<select>`, etc), and prevent focus by [setting their
`[tabindex="-1"]`][mdn-tabindex-1]. Then, re-set the focused element's
`[tabindex="0"]` to treat it as the place to resume once focus is
regained.

To go along with that restriction, treat the up and down arrows as
keyboard shortcuts to move between focusable items in the list.

[mdn-tabindex]: https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/tabindex 93b5eef2be8bd92d6e042afc27726a55bf6f9e80